add agent tests and add limit to torrents per scraper

+++ src/agent.ts
@@ -7,11 +7,13 @@ interface AgentOpts {
   /**
    * Max queries that run at the same time.
    */
-  maxQueries: number;
+  QueriesConcurrency: number;
 }
 
 interface AgentEvents {
   query: [query: Query];
+  query_done: [];
+  destroyed: [];
 }
 
 interface AddQueryOpts {
@@ -20,17 +22,34 @@ interface AddQueryOpts {
   options?: QueryOpts;
 }
 
+export class AgentError extends Error {
+  constructor(msg: string) {
+    super(msg);
+  }
+}
+
 export default class TorrentAgent extends EventEmitter<AgentEvents> {
-  protected queue: PQueue;
+  protected queue: PQueue | null;
+  protected isDestroyed: boolean;
   constructor(opts: Partial<AgentOpts> = {}) {
     super();
-    this.queue = new PQueue({ concurrency: opts.maxQueries || Infinity });
+    this.queue = new PQueue({
+      concurrency: opts.QueriesConcurrency || Infinity,
+    });
+    this.isDestroyed = false;
   }
   add(opts: AddQueryOpts): Query {
+    if (this.isDestroyed) {
+      throw new AgentError("agent is destroyed cannot add new query");
+    }
+    if (!this.queue) {
+      throw new AgentError("queue is destroyed cannot add a new query");
+    }
     const query = new Query(opts.searchQuery, opts.scrapers, opts.options);
     this.queue.add(async () => {
       await query.run();
       await query.destroy();
+      this.emit("query_done");
     });
     this.emit("query", query);
     return query;
@@ -39,6 +58,23 @@ export default class TorrentAgent extends EventEmitter<AgentEvents> {
    * Clear the queue.
    */
   clear() {
+    if (!this.queue) return;
     this.queue.clear();
+  }
+  async destroy() {
+    if (this.isDestroyed) return;
+    this.isDestroyed = true;
+    if (this.queue) {
+      this.queue.pause();
+      this.queue.clear();
+      await this.queue.onIdle();
+      this.queue = null;
+    }
+    this.emit("destroyed");
+    this.removeAllListeners();
+  }
+  // Resolve when the agent is idling
+  async onIdle() {
+    return await this.queue?.onIdle();
   }
 }
